requirements.txt
streamlit
beautifulsoup4
requests
duckduckgo-search==6.2.5
vaderSentiment
python-dateutil
pandas
